1 create data service to save data

+++ src/main/resources/schema-all.sql
@@ -1,1 +1,1 @@
-CREATE TABLE IF NOT EXIST category (code NUMERIC(03) PRIMARY KEY,  categoryText VARCHAR(200), displayLevel NUMERIC(2), selectable VARCHAR(1)  );
+CREATE TABLE IF NOT EXIST category (code serial PRIMARY KEY,  categoryText VARCHAR(200), displayLevel NUMERIC(2), selectable VARCHAR(1)  );
